pinctrl-upboard: add regmap patches for CREX and CRST02 UP Core carrier boards
The initial pinctrl configuration on the CPLD on CREX and CRST02 boards isn't
set up on the BIOS for the pin functions such as SPI, PWM, UART, etc.  So
those functions aren't currently working after boot.

This commit adds a regmap patch, similar to those used on UP/UP^2, to initialise
the pins for the corresponding functions by default.  The GPIO driver will
update the configuration if they are later exported as GPIO pins.

Signed-off-by: Dan O'Donovan <dan@emutex.com>