UPBOARD: Release 4.15.0-37.40~upboard04
Updated release version for Ubuntu 18.04 "Bionic"

NOTE: this change was made on 12th Feb 2019 in a new update pushed to the
Ubuntu PPA, but was never committed here in git until now.  This change
just incremented the version number but unfortunately also changed the
track from Xenial to Bionic.  This is not ideal because we need to
increment the version number again for the next Xenial update, and again
for Bionic, so we end up flipping back and forth.  But we are where we are.
So, for now, lets adopt a convention of using odd-number versions for
Xenial and even-number versions for Bionic.

Signed-off-by: Dan O'Donovan <dan@emutex.com>